requirements.txt
web3>=7.0.0
watchdog>=4.0.0
python-dotenv>=1.0.0
